fix the audience page

+++ lib/weixin/draw_pic.py
@@ -46,8 +46,8 @@ def draw(low_img, headimg, author_name, goal_create_time, goal_content, penalty,
     w, h = content_font.getsize(goal_content)
 
     if len(goal_content) > 6:
-        goal_content1 = goal_content[0: len(goal_content)/2]
-        goal_content2 = goal_content[len(goal_content) / 2:]
+        goal_content1 = goal_content[0: int(len(goal_content)/2)]
+        goal_content2 = goal_content[int(len(goal_content) / 2):]
         w1, h1 = content_font.getsize(goal_content1)
         w2, h2 = content_font.getsize(goal_content2)
